requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.7
django-cors-headers 4.4.0
djangorestframework 3.15.2
mysqlclient         2.2.4
pip                 23.2.1
sqlparse            0.5.1
tzdata              2024.1
